test6

+++ modules/descriptors/include/ssiglib/descriptors/co_occurrence_general.hpp
@@ -45,7 +45,7 @@
 #define COOCALLDIRS
 #define FEATURESCG 12
 #define SIGN(x, y) ((y) < 0 ? -fabs(x) : fabs(x))
-#define SWAP(a, b) {y=(a); (a)=(b); (b)=y;}
+#define SWAP(a, b) {y = (a); (a) = (b); (b) = y;}
 #define PGM_MAXMAXVAL 257
 #define RADIX 2.0
 #define EPSILON 0.000000001
@@ -58,7 +58,7 @@
 #include <opencv2/imgproc.hpp>
 #include "descriptor_2d.hpp"
 
-typedef short gray;
+typedef int gray;
 
 namespace ssig {
 class CoOccurrenceGeneral {
